Skip generated tests for google_tpu_v2_vm (#9315) [upstream:5dc9f2ea88610c0d2a2a94d04548d60e90e1cc54]
